Change to own abstract for cache part instead of magicx
