+++ eZ/Publish/Core/Persistence/Cache/AbstractInMemoryHandler.php
@@ -8,12 +8,13 @@ namespace eZ\Publish\Core\Persistence\Cache;
 
 use eZ\Publish\Core\Persistence\Cache\Adapter\InMemoryClearingProxyAdapter;
 use eZ\Publish\Core\Persistence\Cache\InMemory\InMemoryCache;
-use eZ\Publish\SPI\Persistence\Handler as PersistenceHandler;
 
 /**
  * Class AbstractInMemoryHandler.
  *
- * Abstract handler for use in other Persistence Cache Handlers.
+ * Abstract handler for use in other SPI Handlers.
+ *
+ * @internal
  */
 abstract class AbstractInMemoryHandler
 {
@@ -25,16 +26,14 @@ abstract class AbstractInMemoryHandler
     protected $cache;
 
     /**
-     * @var \eZ\Publish\SPI\Persistence\Handler
-     */
-    protected $persistenceHandler;
-
-    /**
      * @var \eZ\Publish\Core\Persistence\Cache\PersistenceLogger
      */
     protected $logger;
 
     /**
+     * NOTE: On purpose private as it's only supposed to be interacted with in tandem with symfony cache here,
+     *       hence the cache decorator and the reusable methods here.
+     *
      * @var \eZ\Publish\Core\Persistence\Cache\InMemory\InMemoryCache
      */
     private $inMemory;
@@ -42,37 +41,11 @@ abstract class AbstractInMemoryHandler
     /**
      * Setups current handler with everything needed.
      *
-     * @param \eZ\Publish\SPI\Persistence\Handler $persistenceHandler
      * @param \eZ\Publish\Core\Persistence\Cache\Adapter\InMemoryClearingProxyAdapter $cache
      * @param \eZ\Publish\Core\Persistence\Cache\PersistenceLogger $logger
      * @param \eZ\Publish\Core\Persistence\Cache\InMemory\InMemoryCache $inMemory
      */
     public function __construct(
-        PersistenceHandler $persistenceHandler,
-        InMemoryClearingProxyAdapter $cache,
-        PersistenceLogger $logger,
-        InMemoryCache $inMemory
-    ) {
-        $this->persistenceHandler = $persistenceHandler;
-        $this->setCacheDependencies($cache, $logger, $inMemory);
-        $this->init();
-    }
-
-    /**
-     * @api May be used from handlers that needs to inject something else then PersistenceHandler to __construct(), to
-     *      avoid having to have intimate knowledge of internal classes used by AbstractInMemoryHandler in signature.
-     *
-     * E.g:
-     *    public function __construct(MyHandler $myHandler, ...$params) {
-     *        $this->myHandler = $myHandler;
-     *        $this->setCacheDependencies(...$params);
-     *    }
-     *
-     * @param \eZ\Publish\Core\Persistence\Cache\Adapter\InMemoryClearingProxyAdapter $cache
-     * @param \eZ\Publish\Core\Persistence\Cache\PersistenceLogger $logger
-     * @param \eZ\Publish\Core\Persistence\Cache\InMemory\InMemoryCache $inMemory
-     */
-    protected function setCacheDependencies(
         InMemoryClearingProxyAdapter $cache,
         PersistenceLogger $logger,
         InMemoryCache $inMemory
@@ -80,14 +53,6 @@ abstract class AbstractInMemoryHandler
         $this->cache = $cache;
         $this->logger = $logger;
         $this->inMemory = $inMemory;
-    }
-
-    /**
-     * Optional function to initialize handler without having to overload __construct().
-     */
-    protected function init(): void
-    {
-        // overload to add init logic if needed in handler
     }
 
     /**
