+++ eZ/Publish/Core/Persistence/Cache/UserPreferenceHandler.php
@@ -19,7 +19,7 @@ use eZ\Publish\SPI\Persistence\UserPreference\UserPreference;
  *
  * @see \eZ\Publish\SPI\Persistence\UserPreference\Handler
  */
-class UserPreferenceHandler extends AbstractInMemoryHandler implements Handler
+class UserPreferenceHandler extends AbstractInMemoryPersistenceHandler implements Handler
 {
     /**
      * Constant used for storing not found results for getUserPreferenceByUserIdAndName().
